不同权限登录不同页面。

+++ intent-rpc-model/src/main/java/cool/zzy/sems/context/dto/DeliveryDTO.java
@@ -24,5 +24,5 @@ public class DeliveryDTO implements Serializable {
     private String phone;
     private String deliveryName;
     private String avatarUrl;
-    private Integer deliveryStatus;
+    private Short deliveryStatus;
 }
